+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - `cxpc-default-1y`: Fix the 1y chart showing only 30d of data if opened not from `CXM`
+- Fix the `XIT` command with no parameters breaking all later `XIT` commands
 
 ## 25.12.28